fix: check if the form needs to be submitted
The submit `Event` was only triggered, making it possible to execute listeners to this event. However, the form submission itself was never executed (unless manually triggered in an event listener).

This PR fixes checks if the form is supposed to be submitted and does submit it if `e.preventDefault()` was _not_ called.